simulink lineaire fonctionnel

+++ equilibre.m
@@ -29,6 +29,6 @@ eqFC = sign(IBeq)*(ICeq^2 + be1*abs(ICeq))/(ae0 + ZCeq*ae1 + ZCeq^2*ae2 + ZCeq^3
 
 [IAeq IBeq ICeq] = solve([eqFA eqFB eqFC]);
 
-VAeq = IAeq/RA
-VBeq = IBeq/RB
-VCeq = ICeq/RC
+VAeq = double(IAeq/RA)
+VBeq = double(IBeq/RB)
+VCeq = double(ICeq/RC)
